Made Matrix class support parameters which needs to be cast
Before this change the matrix class would give an compile error if it was initialized with parameters of a different type that the matrix. This change allows the parameters to be cast to the internal type of the matrix. This does however also have the downside of silencing any cast warnings that the user may want to receive.

Further work should be done to see if this can be avoided.